backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
bcrypt==4.1.3
itsdangerous==2.2.0
jsonschema==4.23.0
openai==1.40.6
cryptography==43.0.1